added metric calculations

+++ installation/metrics/uspto_metrics_mysql.sql
@@ -16,8 +16,8 @@ CREATE TABLE IF NOT EXISTS uspto.METRICS_G (
   `GrantID` VARCHAR(20) NOT NULL,
   `ForwardCitCnt` INT DEFAULT NULL,
   `BackwardCitCnt` INT DEFAULT NULL,
-  `TCT` INT DEFAULT NULL,
-  `FamilySize` INT DEFAULT NULL,
+  `TCT` FLOAT DEFAULT NULL,
+  `FamilySize` FLOAT DEFAULT NULL,
   PRIMARY KEY (`GrantID`))
   ENGINE = InnoDB;
 
@@ -25,6 +25,7 @@ CREATE TABLE IF NOT EXISTS uspto.METRICS_G (
 -- Create metrics Record for all GrantID
 --
 INSERT INTO uspto.METRICS_G (`GrantID`) SELECT `GrantID` FROM uspto.GRANT;
+
 --
 -- Calculate Forward Citations for all GrantID
 --
@@ -32,12 +33,14 @@ UPDATE uspto.METRICS_G, (select CitedID, count(*) AS count
 FROM uspto.GRACIT_G GROUP BY CitedID) AS t2
 SET uspto.METRICS_G.ForwardCitCnt = t2.count
 WHERE uspto.METRICS_G.GrantID = t2.CitedID;
+
 --
 -- Set all patents without FWD citations to 0
 --
 UPDATE uspto.METRICS_G
 SET uspto.METRICS_G.ForwardCitCnt = 0
 WHERE uspto.METRICS_G.ForwardCitCnt IS NULL;
+
 --
 -- Calculate Backward Citations for all GrantID
 --
@@ -45,6 +48,7 @@ UPDATE uspto.METRICS_G, (select GrantID, count(*) AS count
 FROM uspto.GRACIT_G GROUP BY GrantID) AS t2
 SET uspto.METRICS_G.BackwardCitCnt = t2.count
 WHERE uspto.METRICS_G.GrantID = t2.GrantID;
+
 --
 -- Set all patents without BWD citations to 0
 --
